+++ .idea/misc.xml
@@ -9,12 +9,12 @@
         <entry key="..\:/Kotlin Apps/TriviaGame/app/src/main/res/layout/item_results.xml" value="0.3759057971014493" />
         <entry key="..\:/Kotlin Apps/TriviaGame/app/src/main/res/layout/question_fragment.xml" value="0.25" />
         <entry key="..\:/Kotlin Apps/TriviaGame/app/src/main/res/layout/setup_fragment.xml" value="0.33" />
-        <entry key="..\:/Users/Akram/StudioProjects/Trivia_Game/app/src/main/res/layout/question_fragment.xml" value="0.5" />
-        <entry key="..\:/Users/Akram/StudioProjects/Trivia_Game/app/src/main/res/layout/setup_fragment.xml" value="0.25" />
+        <entry key="..\:/Users/Wesam Nabeel/AndroidStudioProjects/Trivia_Game/app/src/main/res/layout/question_fragment.xml" value="0.128125" />
+        <entry key="..\:/Users/Wesam Nabeel/AndroidStudioProjects/Trivia_Game/app/src/main/res/layout/setup_fragment.xml" value="0.16614583333333333" />
       </map>
     </option>
   </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" project-jdk-name="1.8" project-jdk-type="JavaSDK">
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="1.8" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
   </component>
   <component name="ProjectType">
