+++ app/src/main/java/com/example/triviagame/data/domain/data/QuestionsAnswers.kt
@@ -1,7 +1,6 @@
 package com.example.triviagame.data.domain.data
 
 
-import com.example.triviagame.data.domain.Results
 import com.google.gson.annotations.SerializedName
 
 data class QuestionsAnswers (
