+++ app/src/main/java/com/example/triviagame/data/domain/data/TriviaResponse.kt
@@ -1,5 +1,6 @@
-package com.example.triviagame.data.domain
+package com.example.triviagame.data.domain.data
 
+import com.example.triviagame.data.domain.data.Results
 import com.google.gson.annotations.SerializedName
 
 data class TriviaResponse (
