+++ app/src/main/java/com/example/triviagame/data/domain/data/TriviaResponseMain.kt
@@ -1,11 +1,12 @@
-package com.example.triviagame.data.domain
+package com.example.triviagame.data.domain.data
 
 
+import com.example.triviagame.data.domain.data.Results
 import com.google.gson.annotations.SerializedName
 
 data class TriviaResponseMain(
     @SerializedName("response_code")
     val responseCode: Int?,
     @SerializedName("results")
-    val results: List<Result>?
+    val results: List<Results>?
 )
