+++ app/src/main/java/com/example/triviagame/extensions.kt
@@ -6,3 +6,4 @@ import io.reactivex.rxjava3.disposables.Disposable
 fun Disposable.addTo(compositeDisposable: CompositeDisposable){
     compositeDisposable.add(this)
 }
+
